setup react native router
